test: split luxo lamp integration suite by computation for CI shard balance
luxoLampClevis.test.ts was the slowest single test file (399s), making it
the wall-clock floor of the per-file CI shard matrix. Restructure without
changing what is tested:

- luxoLampClevis.validate.test.ts: the three kinematic-loop assertions
  each re-ran runValidateCli with the identical input (same file,
  epsilon: 0.01, includeInterference: true, physical: false, json: true,
  includePhysics: false). The run is now computed once in beforeAll and
  the tests assert against the shared result — same assertions, same
  thresholds, same inputs (dedup: 3 identical validate runs -> 1).
- The example sweep gate's per-example check for luxo-lamp ran a fourth
  identical validate; that test moved into the same file to share the
  beforeAll run. exampleSweepGate.test.ts delegates it via the new
  HOSTED_IN_DEDICATED_FILE registry and structurally asserts the hosting
  file exists and references the example, preserving silent-skip
  resistance.
- luxoLampClevis.mjcf.test.ts: the P11 MJCF emission test is a different
  computation (runScript + assemblyToMjcf) and pays its own OCCT init.
- luxoLampClevis.test.ts keeps the two source-regex tests and the
  issue-379 it.skip unchanged.

Measured per-file: 399s -> 97s (validate, single irreducible validate
run) + 22s (mjcf) + 2s (source); sweep gate 193s -> 16s. Test counts
unchanged: 7 + 37 -> 3 + 4 + 1 + 36.
