+++ tests/integration/examples/luxoLampClevis.test.ts
@@ -15,15 +15,17 @@
 //
 // Spec:  docs/specs/2026-06-01-physics-grounded-loop-design.md
 // Plan:  docs/plans/2026-06-01-physics-loop-P5-luxo-geometric-rebuild.md
+//
+// NOTE: companion files were split out for CI shard balance (per-file
+// vitest sharding):
+//   - luxoLampClevis.validate.test.ts — the three kinematic-loop
+//     validate assertions (shared single `runValidateCli` run) plus the
+//     example-sweep-gate check for this example.
+//   - luxoLampClevis.mjcf.test.ts — the P11 Slice 1 MJCF emission test.
 
 import { describe, expect, it } from 'vitest';
 import { readFileSync } from 'node:fs';
-import { dirname, resolve } from 'node:path';
 import { runValidateCli } from '../../../src/agent/cli/commands/validate';
-import { initOcct } from '../../../src/kernel/backends/occt/occtBackend';
-import { runScript } from '../../../src/modeling/runtime/runScript';
-import { assemblyToMjcf } from '../../../src/modeling/runtime/mjcfExport';
-import type { Assembly } from '../../../src/modeling/capture/assembly';
 
 const LUXO_SCRIPT_PATH = 'examples/kinematic/luxo-lamp.kcad.ts';
 
@@ -48,104 +50,6 @@ describe('Luxo lamp — passes the physics-grounded loop', () => {
     const revoluteMates = source.match(/['"]revolute['"]/g) ?? [];
     expect(revoluteMates.length).toBeGreaterThanOrEqual(3);
   });
-
-  it('passes the kinematic-only mechanism-truth loop incl. criterion 8: mechanism: real (P11 Slice 3 wrap re-author)', async () => {
-    // P9 made the lamp pass criteria 1-4 + 7. P11 Slice 2 added criterion
-    // 8 (tendon-body-intersect), which flagged the straight springs
-    // cutting through the arms. P11 Slice 3 routes each spring over a wrap
-    // rail on the beam it spans, so no cable pierces a body — the lamp is
-    // back to `mechanism: real` with the spring-through-structure bug
-    // closed.
-    const r = await runValidateCli({
-      file: LUXO_SCRIPT_PATH,
-      epsilon: 0.01,
-      includeInterference: true,
-      physical: false,
-      json: true,
-      includePhysics: false,
-    });
-    expect(r.mechanism).toBe('real');
-    expect(r.mechanismFailures ?? []).toEqual([]);
-  }, 240_000);
-
-  it('is interference-clean across the full pose sweep: ZERO mechanism.interpenetration diagnostics (2026-06-03 absolute 20 mm³ gate)', async () => {
-    // The assertion that was MISSING and let the slop through (spec
-    // sequencing step 6): under the absolute 20 mm³ interference gate the
-    // lamp must have NO interpenetration at any sampled pose. The pre-redesign
-    // 5%-of-bbox / 2500 mm³-revolute caps hid body-near-fork overlaps; the
-    // clevis tongue-drill (decision #2), the head neck-stem restructure, and
-    // the tightened elbow (-95°) / wrist (-50°) limits make every sampled pose
-    // interference-clean. This is the `interferences === 0` contract.
-    const r = await runValidateCli({
-      file: LUXO_SCRIPT_PATH,
-      epsilon: 0.01,
-      includeInterference: true,
-      physical: false,
-      json: true,
-      includePhysics: false,
-    });
-    const interferences = (r.mechanismFailures ?? []).filter(
-      (f) => f.code === 'mechanism.interpenetration',
-    );
-    expect(interferences).toEqual([]);
-    expect(r.mechanism).toBe('real');
-  }, 240_000);
-
-  it('P8 joint-mesh-continuity gate sees NO joint-mesh-gap diagnostics on the post-P9 Luxo (GREEN after P9)', async () => {
-    // P9 (2026-06-02): with the spring-boss posts on each arm and the
-    // extended column / pulled-back head-neck, every mate connector
-    // lies inside its body's mesh within 1 mm — no
-    // `mechanism.joint-mesh-gap` diagnostics fire.
-    const r = await runValidateCli({
-      file: LUXO_SCRIPT_PATH,
-      epsilon: 0.01,
-      includeInterference: true,
-      physical: false,
-      json: true,
-      includePhysics: false,
-    });
-    const gaps = (r.mechanismFailures ?? []).filter(
-      (f) => f.code === 'mechanism.joint-mesh-gap',
-    );
-    expect(gaps).toEqual([]);
-    expect(r.mechanism).toBe('real');
-  }, 240_000);
-
-  it('P11 Slice 1: emitted MJCF contains one <mesh> + one <geom type="mesh"> per part', async () => {
-    // The lamp builds 4 parts via `arm.part('base' | 'lower-arm' |
-    // 'upper-arm' | 'lamp-head', ...)`. Post-Slice 1, every part
-    // contributes one inline `<asset><mesh>` and one
-    // `<geom type="mesh">` inside its `<body>` — MuJoCo's drop-test
-    // becomes collision-aware. Asserting the counts guards against a
-    // regression that silently drops the geom emit for one body.
-    await initOcct();
-    const absPath = resolve(LUXO_SCRIPT_PATH);
-    const code = readFileSync(absPath, 'utf8');
-    const run = await runScript({
-      code,
-      fileName: LUXO_SCRIPT_PATH,
-      scriptDir: dirname(absPath),
-    });
-    const arms = Array.from(run.session.assemblies.values()) as readonly Assembly[];
-    expect(arms.length).toBeGreaterThan(0);
-    const lamp = arms[0];
-    const { mjcf } = await assemblyToMjcf(lamp);
-    const meshAssetCount = (mjcf.match(/<mesh name="part-/g) ?? []).length;
-    const geomMeshCount = (mjcf.match(/<geom type="mesh"/g) ?? []).length;
-    expect(meshAssetCount).toBe(4);
-    expect(geomMeshCount).toBe(4);
-    expect(mjcf).toMatch(/<size nconmax="500"\/>/);
-    // P11 Slice 1: every mate emits one `<contact><exclude>`. The
-    // lamp has 3 revolute mates (shoulder, elbow, wrist) — without the
-    // excludes, the clevis fork-tongue mesh overlap at every joint
-    // would generate spurious contact forces strong enough to fling
-    // the chain apart in the drop-test integrator. (Tendon emission is
-    // separate from the mate count; tendons emit `<spatial>`, not
-    // `<exclude>`.)
-    expect(mjcf).toMatch(/<contact>/);
-    const excludeCount = (mjcf.match(/<exclude /g) ?? []).length;
-    expect(excludeCount).toBe(3);
-  }, 240_000);
 
   // SKIPPED — tracked under issues/379 (PHYSICS spring re-calibration, NOT an
   // interference issue). The mechanism-validity absolute-cap redesign
